requirements.txt
flask
pydantic
python-dotenv
google-adk
google-genai
pillow
reportlab
